# openapi-3.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jiro Search API",
    "version": "0.2.0",
    "description": "Local-first web search, scraping, and social media intelligence platform. Supports 9+ search engines, 12 social platforms, hybrid search, structured extraction, and smart intent routing.",
    "contact": {
      "name": "Jiro Team",
      "url": "https://github.com/DevAnimecx/jiro"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.jiro.dev",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/v1/social": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Scrape social media URL",
        "description": "Scrape a social media post, profile, or thread from 12 supported platforms.",
        "operationId": "socialScrape",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialScrapeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Social content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported platform"
          }
        }
      }
    },
    "/v1/social/batch": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Batch scrape social URLs",
        "description": "Scrape multiple social media URLs in parallel.",
        "operationId": "socialBatch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialBatchResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/social/search": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Search social platforms",
        "description": "Search across multiple social platforms simultaneously.",
        "operationId": "socialSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialSearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/social/search/everywhere": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Search all social platforms",
        "description": "Search across all 12 supported social platforms.",
        "operationId": "socialSearchEverywhere",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialSearchEverywhereRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialSearchResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}